+++ include/electionguard/ballot_compact.h
@@ -43,7 +43,7 @@ EG_API eg_electionguard_status_t eg_compact_plaintext_ballot_from_msgpack(
   uint8_t *in_data, uint64_t in_length, eg_compact_plaintext_ballot_t **out_handle);
 
 EG_API eg_electionguard_status_t eg_compact_plaintext_ballot_to_msgpack(
-  eg_compact_plaintext_ballot_t *handle, uint8_t **out_data, size_t *out_size);
+  eg_compact_plaintext_ballot_t *handle, uint8_t **out_data, uint64_t *out_size);
 
 EG_API eg_electionguard_status_t eg_compact_plaintext_ballot_msgpack_free(uint8_t *in_data);
 
@@ -78,7 +78,7 @@ EG_API eg_electionguard_status_t eg_compact_ciphertext_ballot_from_msgpack(
   uint8_t *in_data, uint64_t in_length, eg_compact_ciphertext_ballot_t **out_handle);
 
 EG_API eg_electionguard_status_t eg_compact_ciphertext_ballot_to_msgpack(
-  eg_compact_ciphertext_ballot_t *handle, uint8_t **out_data, size_t *out_size);
+  eg_compact_ciphertext_ballot_t *handle, uint8_t **out_data, uint64_t *out_size);
 
 EG_API eg_electionguard_status_t eg_compact_ciphertext_ballot_msgpack_free(uint8_t *in_data);
