+++ include/electionguard/hash.h
@@ -11,7 +11,7 @@ extern "C" {
 
 EG_API eg_electionguard_status_t eg_hash_elems_string(const char *a,
                                                       eg_element_mod_q_t **out_handle);
-EG_API eg_electionguard_status_t eg_hash_elems_strings(const char **a, size_t in_length,
+EG_API eg_electionguard_status_t eg_hash_elems_strings(const char **a, uint64_t in_length,
                                                        eg_element_mod_q_t **out_handle);
 EG_API eg_electionguard_status_t eg_hash_elems_int(const uint64_t a,
                                                    eg_element_mod_q_t **out_handle);
